pyproject.toml
[project]
name = "pysystemtradeoncontainers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.27.2",
    "orjson>=3.10.7",
    "pandas>=2.2.2",
    "pandera[mypy]>=0.20.4",
    "pydantic-settings>=2.5.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    "redis>=5.0.8",
    "rich>=13.8.1",
]
[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --cov=risk --cov=common --cov-report=term-missing:skip-covered"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 140
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    'A',     # flake8-builtins
    'ASYNC', # flake8-async
    'B',     # flake8-bugbear
    'C90',   # mccabe
    'COM',   # flake8-commas (flake8-simplify)
    'E',     # pycodestyle errors
    'ERA',   # eradicate dead code
    'EXE',   # flake8-executable
    'F',     # Pyflakes
    'FA',    # flake8-annotations
    'FLY',   # flynt
    'G',     # flake8-logging-format
    'I',     # isort
    'INP',   # flake8-no-pep420
    'INT',   # flake8-gettext
    'ISC',   # flake8-implicit-str-concat
    'LOG',   # flake8-logging
    'PERF',  # Perflint
    'PIE',   # flake8-pie
    'PL',    # Pylint
    'PLR',
    'PTH',   # flake8-use-pathlib
    'Q',     # flake8-quotes
    'RET',   # flake8-return
    'RSE',   # flake8-raise
    'RUF',   # Ruff-specific rules
    'S',     # flake8-bandit (security)
    'SIM',   # flake8-simplify
    'SLF',   # flake8-self
    'T20',   # flake8-print
    'TCH',   # flake8-type-checking
    'TID',   # flake8-tidy-imports
    'TRY',   # tryceratops
    'UP',    # pyupgrade
    'W',     # pycodestyle warnings
    'YTT',   # flake8-2020
]

ignore = [
    'S104',   # not a issue in docker
    'TRY003', # i do not like this ... it makes validation harder
    'TRY300', # rarely used in python and is not know by people
    'TRY301', #fastapi recommended way to raise exception
    'COM812', #conflict with formatter
    'ISC001', #conflict with formatter
    'RUF012', # not compatible with beanie
]

exclude = [
    ".git",
    ".mypy_cache",
    ".pyenv",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "venv",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Form", "fastapi.File"]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.mypy]
plugins = ["pydantic.mypy", "pandera.mypy"]
ignore_missing_imports = true
no_site_packages = true


follow_imports = 'silent'
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
